support char.bin
