+++ momiji-core/build.gradle.kts
@@ -22,7 +22,7 @@ kotlin {
         }
     }
     macosArm64()
-//    macosX64()
+    macosX64()
     linuxX64()
 
     sourceSets {
