+++ momiji-core/src/jvmTest/kotlin/io/github/tokuhirom/momiji/core/character/CharMapJVMTest.kt
@@ -1,4 +1,4 @@
-package io.github.tokuhirom.momiji.core.src
+package io.github.tokuhirom.momiji.core.character
 
 import okio.FileSystem
 import okio.Path.Companion.toPath
@@ -7,6 +7,23 @@ import kotlin.test.assertEquals
 
 class CharMapJVMTest {
     private val fileSystem = FileSystem.SYSTEM
+
+    @Test
+    fun testText() {
+        val path = "../".toPath().resolve("momiji-ipadic/build/dict/char.def")
+        if (!fileSystem.exists(path)) {
+            println("file not found. skip the test.: $path")
+            return
+        }
+
+        val charMap =
+            CharMap.parseText(
+                fileSystem.read(path) {
+                    this.readUtf8()
+                },
+            )
+        assertEquals("ALPHA", charMap.categoryName(charMap.resolve('A').defaultType))
+    }
 
     @Test
     fun testBinary() {
